refactor: move useDebauncedValue to hooks
